Still wip, adding some shit

+++ functions.h
@@ -8,14 +8,15 @@
 
 namespace neuro {
 	
-	using Activation = std::function< double(double, double) >;
+	using Activation = std::function< float(float, float) >;
 	
 	std::map< std::string, Activation > activations = 
 	{
-	{ "linear", 	[](double x, double a) { return x*a; } }, 
-	{ "tanh",		[](double x, double a) { return activations["sigmoid"](x,a)*2-1 ;} }, 
-	{ "sigmoid", 	[](double x, double a) { return 1/(1+exp(-x*a)); } },
-	{ "gauss",		[](double x, double a) { return ( 1/(exp(a)*sqrt(std::numbers::pi*2)) )*exp( -0.5*std::pow( x/exp(a), 2 ) ); 	} 	}, //Sigma is exp of parameter - in order to prevent div by zero.
+	{ "linear", 	[](float x, float a) { return x*a; } }, 
+	{ "tanh",		[](float x, float a) { return activations["sigmoid"](x,a)*2-1 ;} }, 
+	{ "sigmoid", 	[](float x, float a) { return 1/(1+exp(-x*a)); } },
+// TODO - gauss is not working
+	{ "gauss",		[](float x, float a) { return ( 1/(exp(a)*sqrt(std::numbers::pi*2)) )*exp( -0.5*std::pow( x/exp(a), 2 ) ); 	} 	}, //Sigma is exp of parameter - in order to prevent div by zero.
 	
 // Logistic sigmoid
 // Arctg
@@ -26,8 +27,8 @@ namespace neuro {
 // Leaky ReLU
 // Спизди с википедии из статьи "Функции активации"
 
-	{ "treshold",	[](double x, double a) { if(x>a) return 1; else return 0; } },	
-//	{ "name",	[](double x, double a) { return 0;} },  // TEMPLATE
+	{ "treshold",	[](float x, float a) { if(x>a) return 1; else return 0; } },	
+//	{ "name",	[](float x, float a) { return 0;} },  // TEMPLATE
 	};
